chore(refactor): organize the code into multiple modules
